Inaccurate calculate of the ping
Sending a long and unnecessary data snippet like 'ping' will only increase the delay. Here we need to measure the latency between Discord and the Bot, not the long message sending. So the smaller the length of the outgoing packet, the more accurate data we get. I have used 0 for now, but we can change it in the future.